Validação erro

+++ src/main/java/com/devsuperior/descommerce/services/ProductService.java
@@ -1,16 +1,20 @@
 package com.devsuperior.descommerce.services;
 
-import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devsuperior.descommerce.dto.ProductDTO;
 import com.devsuperior.descommerce.entities.Product;
 import com.devsuperior.descommerce.repositories.ProductRepository;
+import com.devsuperior.descommerce.services.exceptions.DataBaseExeption;
+import com.devsuperior.descommerce.services.exceptions.ResourceNotFoundExeption;
+
+import jakarta.persistence.EntityNotFoundException;
 
 
 
@@ -23,7 +27,8 @@ public class ProductService {
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
 		
-		Product entity = repository.findById(id).get();
+		Product entity = repository.findById(id).orElseThrow(
+				() -> new ResourceNotFoundExeption("Produto não encontrado"));
 		return new ProductDTO(entity);				
 	}
 	
@@ -48,19 +53,30 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO dto) {
-		
+		try {
 		Product entity = repository.getReferenceById(id); //ainda não foi ao banco de dados, mas é monitorado pela jpa
 		copyDTOtoEntity(dto, entity);		
 		entity = repository.save(entity);		
 		return new ProductDTO(entity);
+		} catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundExeption("Producto para atualizar não existe");
+		}
 				
 	}
 	
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
 		
-		repository.deleteById(id);
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundExeption("Produto para excluir não existe");
+		}
+		
+		try {
+			repository.deleteById(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DataBaseExeption("Violação de integridade referencial");
+		}
 		
 	}
